Communicator/processor/requirements.txt
# FastAPI and HTTP dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
# HTTP Clients
httpx>=0.25.0
requests>=2.31.0
python-multipart>=0.0.6

# gRPC dependencies
grpcio>=1.59.0
grpcio-tools>=1.59.0

# Kafka/messaging dependencies
aiokafka>=0.8.10

# Configuration and environment
python-dotenv>=1.0.0

# Async utilities
asyncio-mqtt>=0.16.1

# Date/time utilities
python-dateutil>=2.8.2

# Logging and monitoring
structlog>=23.2.0

# Redis for caching/queues (optional)
redis>=5.0.0

# Development dependencies (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0